@@ -3,18 +3,17 @@ package pl.zajacp;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import pl.zajacp.model.GameRecord;
 import pl.zajacp.model.GamesLog;
 import pl.zajacp.repository.DynamoDbRepository;
+import pl.zajacp.repository.GamesLogRepository;
 
+@NoArgsConstructor
 @AllArgsConstructor
 public class PutGamesLogHandler implements RequestHandler<GamesLog, String> {
 
-    private DynamoDbRepository<GameRecord> gameItemRepository;
-
-    public PutGamesLogHandler() {
-        gameItemRepository = new DynamoDbRepository<>("games_log", GameRecord.class);
-    }
+    private DynamoDbRepository<GameRecord> gameItemRepository = GamesLogRepository.getInstance();
 
     @Override
     public String handleRequest(GamesLog gamesLog, Context context) {
